+++ test/src/deps_sort/_deps_parser_test.dart
@@ -1,8 +1,8 @@
 import 'dart:io';
 
+import 'package:dart_dependency_checker/src/_shared/package.dart';
 import 'package:dart_dependency_checker/src/dependency_type.dart';
 import 'package:dart_dependency_checker/src/deps_sort/_deps_parser.dart';
-import 'package:dart_dependency_checker/src/deps_sort/model/package.dart';
 import 'package:dart_dependency_checker/src/util/yaml_map_ext.dart';
 import 'package:dart_dependency_checker/src/util/yaml_map_loader.dart';
 import 'package:test/test.dart';
@@ -20,22 +20,17 @@ void main() {
     expect(result, {
       const HostedPackage('meta', '^1.11.0'),
       const HostedPackage('args', '2.4.2'),
-      const HostedPackage(
-        'some_hosted_source',
-        '^1.4.0',
-        hostedUrl: 'https://some-package-server.com',
-      ),
-      const SdkPackage('flutter', sdk: 'flutter'),
-      const PathPackage('some_path_source', path: '../some_path_dependency'),
-      const PathPackage('some_path_source2', path: '../some_path_dependency2'),
+      const SdkPackage('flutter', 'flutter'),
+      const PathPackage('some_path_source', '../some_path_dependency'),
+      const PathPackage('some_path_source2', '../some_path_dependency2'),
       const HostedPackage('equatable', '^2.0.7'),
       const GitPackage(
         'window_size',
-        url: 'git@github.com:google/flutter-desktop-embedding.git',
+        'git@github.com:google/flutter-desktop-embedding.git',
         path: 'plugins/window_size',
         ref: 'e48abe7c3e9ebfe0b81622167c5201d4e783bb81',
       ),
-      const GitPackage('yaansi', url: 'https://github.com/akaiser/yaansi.git'),
+      const GitPackage('yaansi', 'https://github.com/akaiser/yaansi.git'),
       const HostedPackage('path', '^1.9.1'),
       const HostedPackage('yaml', '3.1.3'),
     });